Changed from Working from Threads to Coroutines and started DataRetrieval
Unity does not support working with multiple threads, so the WebListener had to be switched to work with a Coroutine instead when waiting for a packet.

Began working on DataRetrieval to give an idea of how it would work in a real application. Currently, the actual arguments that will be passed via Dictionary are unknown so this class contains only test actions.